requirements.txt
annotated-types==0.7.0
contourpy==1.3.3
coverage==7.13.1
cycler==0.12.1
-e git+https://github.com/MagicSpik3/etl-ir-core@ca98f8fa9bedfc4d1894a772ababff62227bb6c4#egg=etl_ir_core
fonttools==4.61.1
iniconfig==2.3.0
kiwisolver==1.4.9
matplotlib==3.10.8
networkx==3.6.1
numpy==2.4.1
packaging==25.0
pillow==12.1.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
